bills for day and month

+++ src/components/Table/index.js
@@ -1,7 +1,8 @@
 import React from "react";
 import firebase from "firebase";
-import firebaseConfig from "../../ConfigFirebase";
 
+import { Swaling } from "../../Helpers/afterActions";
+import firebaseConfig from "../../ConfigFirebase";
 import TableRow from "./TableRow";
 import TableButton from "./TableButton";
 import { getdata, deleteData } from "../.././Helpers/HandleFirebase";
@@ -26,9 +27,13 @@ class Table extends React.Component {
   };
 
   deleteMember = async key => {
-    let onData = deleteData(key);
-    onData.remove();
-    this.loadData();
+    let comfirm = window.confirm("Are you sure you wish to delete this item?");
+    if (comfirm) {
+      let onData = deleteData(key);
+      onData.remove();
+      Swaling("Information deleted !");
+      this.loadData();
+    }
   };
 
   render() {
